Remove circuits traits
