+++ Cargo.toml
@@ -48,7 +48,6 @@ members = [
   "zk/circuits/prover",
   "zk/circuits/utils",
   "zk/circuits/verifier",
-  "zk/circuits/witness_generators/core",
   "zk/circuits/witness_generators/pol",
   "zk/groth16",
   "zk/poseidon2",
@@ -109,7 +108,6 @@ services-utils           = { default-features = false, path = "./nomos-services/
 subnetworks-assignations = { default-features = false, path = "./nomos-da/network/subnetworks-assignations" }
 tests                    = { default-features = false, path = "./tests" }
 utxotree                 = { default-features = false, path = "./utxotree" }
-witness-generator-core   = { default-features = false, path = "./zk/circuits/witness_generators/core" }
 
 # External
 blake2            = { default-features = false, version = "0.10" }
