+++ zk/circuits/prover/src/lib.rs
@@ -1,6 +1,3 @@
-mod prover;
 mod rapidsnark;
-mod wrappers;
 
-pub use prover::Prover;
-pub use rapidsnark::Rapidsnark;
+pub use rapidsnark::*;
