--- zk/circuits/witness_generators/core/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-edition           = "2024"
-license.workspace = true
-name              = "witness-generator-core"
-version           = "0.1.0"
-
-[dependencies]
-
-[lints]
-workspace = true
